Reimplementing the Value_type, Size_type, and Distance_type frameworks using new deduction rules.
git-svn-id: http://origin.googlecode.com/svn/trunk@437 3dbf15f1-5c06-4b2d-4fb8-2ed049db4e7e

+++ core/tests/concepts/difference_type.cpp
@@ -21,7 +21,7 @@ int main()
 {
   // Check adaptation for some arithmetic types
   static_assert(Has_distance_type<int>(), "");
-  static_assert(Same<Distance_type<int>, ptrdiff_t>(), "");
+  static_assert(Same<Distance_type<int>, int>(), "");
 
   // Check adaptation for pointers.
   static_assert(Has_distance_type<int*>(), "");
